+++ docarray/array/storage/base/backend.py
@@ -1,0 +1,7 @@
+from abc import ABC, abstractmethod
+
+
+class BaseBackendMixin(ABC):
+    @abstractmethod
+    def _init_storage(self, *args, **kwargs):
+        ...
