+++ docarray/array/storage/memory/__init__.py
@@ -2,6 +2,8 @@ from .backend import MemoryBackendMixin
 from .getsetdel import GetSetDelMixin
 from .seqlike import SequenceLikeMixin
 
+__all__ = ['MemoryStorageMixins']
+
 
 class MemoryStorageMixins(MemoryBackendMixin, GetSetDelMixin, SequenceLikeMixin):
     ...
