+++ docarray/array/storage/sqlite/__init__.py
@@ -1,0 +1,9 @@
+from .backend import SqliteBackendMixin
+from .getsetdel import GetSetDelMixin
+from .seqlike import SequenceLikeMixin
+
+__all__ = ['SqliteStorageMixins']
+
+
+class SqliteStorageMixins(SqliteBackendMixin, GetSetDelMixin, SequenceLikeMixin):
+    ...
