+++ docarray/array/storage/sqlite/seqlike.py
@@ -1,0 +1,56 @@
+from typing import Iterator, Union, Sequence, Iterable, MutableSequence
+
+from .... import Document
+
+
+class SequenceLikeMixin(MutableSequence[Document]):
+    """Implement sequence-like methods"""
+
+    def insert(self, index: int, value: 'Document'):
+        """Insert `doc` at `index`.
+
+        :param index: Position of the insertion.
+        :param value: The doc needs to be inserted.
+        """
+        length = len(self)
+        if index < 0:
+            index = length + index
+        index = max(0, min(length, index))
+        self._shift_index_right_backward(index)
+        self._insert_doc_at_idx(doc=value, idx=index)
+        self.connection.commit()
+
+    def append(self, value: 'Document') -> None:
+        self._insert_doc_at_idx(value)
+        self.connection.commit()
+
+    def extend(self, values: Iterable['Document']) -> None:
+        idx = len(self)
+        for v in values:
+            self._insert_doc_at_idx(v, idx)
+            idx += 1
+        self.connection.commit()
+
+    def clear(self) -> None:
+        self._sql(f'DELETE FROM {self.table_name}')
+        self.connection.commit()
+
+    def __contains__(self, item: Union[str, 'Document']):
+        if isinstance(item, str):
+            r = self._sql(f"SELECT 1 FROM {self.table_name} WHERE doc_id=?", (item,))
+            return len(list(r)) > 0
+        elif isinstance(item, Document):
+            return item.id in self  # fall back to str check
+        else:
+            return False
+
+    def __len__(self) -> int:
+        r = self._sql(f'SELECT COUNT(*) FROM {self.table_name}')
+        return r.fetchone()[0]
+
+    def __iter__(self) -> Iterator['Document']:
+        r = self._sql(
+            f'SELECT serialized_value FROM {self.table_name} ORDER BY item_order'
+        )
+        for res in r:
+            yield res[0]
